Move newPrivateKey into YubiKeyService.NewPrivateKey.
